refactor(orulo): extract tags processor
